+++ src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/interpreter/InterpreterSistaV1Node.java
@@ -419,7 +419,7 @@ public final class InterpreterSistaV1Node extends AbstractInterpreterNode {
                     case BC.BYTECODE_PRIM_ADD: {
                         final Object arg = pop(frame, --sp);
                         final Object receiver = popReceiver(frame, --sp);
-                        final byte state = profiles[currentPC];
+                        final short state = profiles[currentPC];
                         final Object result;
                         if (receiver instanceof final Long lhs && arg instanceof final Long rhs) {
                             if ((state & 0b1000) == 0) {
@@ -462,7 +462,7 @@ public final class InterpreterSistaV1Node extends AbstractInterpreterNode {
                     case BC.BYTECODE_PRIM_SUBTRACT: {
                         final Object arg = pop(frame, --sp);
                         final Object receiver = popReceiver(frame, --sp);
-                        final byte state = profiles[currentPC];
+                        final short state = profiles[currentPC];
                         final Object result;
                         if (receiver instanceof final Long lhs && arg instanceof final Long rhs) {
                             if ((state & 0b1000) == 0) {
@@ -505,7 +505,7 @@ public final class InterpreterSistaV1Node extends AbstractInterpreterNode {
                     case BC.BYTECODE_PRIM_LESS_THAN: {
                         final Object arg = pop(frame, --sp);
                         final Object receiver = popReceiver(frame, --sp);
-                        final byte state = profiles[currentPC];
+                        final short state = profiles[currentPC];
                         final Object result;
                         if (receiver instanceof final Long lhs && arg instanceof final Long rhs) {
                             if ((state & 0b1000) == 0) {
@@ -534,7 +534,7 @@ public final class InterpreterSistaV1Node extends AbstractInterpreterNode {
                     case BC.BYTECODE_PRIM_GREATER_THAN: {
                         final Object arg = pop(frame, --sp);
                         final Object receiver = popReceiver(frame, --sp);
-                        final byte state = profiles[currentPC];
+                        final short state = profiles[currentPC];
                         final Object result;
                         if (receiver instanceof final Long lhs && arg instanceof final Long rhs) {
                             if ((state & 0b1000) == 0) {
@@ -563,7 +563,7 @@ public final class InterpreterSistaV1Node extends AbstractInterpreterNode {
                     case BC.BYTECODE_PRIM_LESS_OR_EQUAL: {
                         final Object arg = pop(frame, --sp);
                         final Object receiver = popReceiver(frame, --sp);
-                        final byte state = profiles[currentPC];
+                        final short state = profiles[currentPC];
                         final Object result;
                         if (receiver instanceof final Long lhs && arg instanceof final Long rhs) {
                             if ((state & 0b1000) == 0) {
@@ -592,7 +592,7 @@ public final class InterpreterSistaV1Node extends AbstractInterpreterNode {
                     case BC.BYTECODE_PRIM_GREATER_OR_EQUAL: {
                         final Object arg = pop(frame, --sp);
                         final Object receiver = popReceiver(frame, --sp);
-                        final byte state = profiles[currentPC];
+                        final short state = profiles[currentPC];
                         final Object result;
                         if (receiver instanceof final Long lhs && arg instanceof final Long rhs) {
                             if ((state & 0b1000) == 0) {
@@ -621,7 +621,7 @@ public final class InterpreterSistaV1Node extends AbstractInterpreterNode {
                     case BC.BYTECODE_PRIM_EQUAL: {
                         final Object arg = pop(frame, --sp);
                         final Object receiver = popReceiver(frame, --sp);
-                        final byte state = profiles[currentPC];
+                        final short state = profiles[currentPC];
                         final Object result;
                         if (receiver instanceof final Long lhs && arg instanceof final Long rhs) {
                             if ((state & 0b1000) == 0) {
@@ -650,7 +650,7 @@ public final class InterpreterSistaV1Node extends AbstractInterpreterNode {
                     case BC.BYTECODE_PRIM_NOT_EQUAL: {
                         final Object arg = pop(frame, --sp);
                         final Object receiver = popReceiver(frame, --sp);
-                        final byte state = profiles[currentPC];
+                        final short state = profiles[currentPC];
                         final Object result;
                         if (receiver instanceof final Long lhs && arg instanceof final Long rhs) {
                             if ((state & 0b1000) == 0) {
@@ -679,7 +679,7 @@ public final class InterpreterSistaV1Node extends AbstractInterpreterNode {
                     case BC.BYTECODE_PRIM_BIT_AND: {
                         final Object arg = pop(frame, --sp);
                         final Object receiver = popReceiver(frame, --sp);
-                        final byte state = profiles[currentPC];
+                        final short state = profiles[currentPC];
                         final Object result;
                         if (receiver instanceof final Long lhs && arg instanceof final Long rhs) {
                             if ((state & 0b1000) == 0) {
@@ -702,7 +702,7 @@ public final class InterpreterSistaV1Node extends AbstractInterpreterNode {
                     case BC.BYTECODE_PRIM_BIT_OR: {
                         final Object arg = pop(frame, --sp);
                         final Object receiver = popReceiver(frame, --sp);
-                        final byte state = profiles[currentPC];
+                        final short state = profiles[currentPC];
                         final Object result;
                         if (receiver instanceof final Long lhs && arg instanceof final Long rhs) {
                             if ((state & 0b1000) == 0) {
@@ -724,7 +724,7 @@ public final class InterpreterSistaV1Node extends AbstractInterpreterNode {
                     }
                     case BC.BYTECODE_PRIM_SIZE: {
                         final Object receiver = popReceiver(frame, --sp);
-                        final byte state = profiles[currentPC];
+                        final short state = profiles[currentPC];
                         final Object result;
                         if (receiver instanceof final NativeObject string && image.isByteString(string)) {
                             if ((state & 0b1000) == 0) {
